+++ src/nodes/NodeAddressContent.h
@@ -8,31 +8,30 @@
 #ifndef NODEADDRESSCONTENT_H_
 #define NODEADDRESSCONTENT_H_
 
-template <unsigned int DIM>
+template<unsigned int DIM>
 class Node;
 
-template <unsigned int DIM>
+template<unsigned int DIM>
 struct NodeAddressContent {
-	// TODO reduce bool values or align them!
-	bool exists;
-	bool hasSubnode;
-	bool directlyStoredSuffix;
-	bool hasSpecialPointer;
-	uint64_t id;
-	unsigned long address;
-	union {
-		Node<DIM>* subnode;
-		const unsigned long* suffixStartBlock;
-		unsigned int suffixStartBlockIndex;
-		unsigned long suffix;
-		uintptr_t specialPointer;
-	};
+  // TODO reduce bool values or align them!
+  bool exists;
+  bool hasSubnode;
+  bool directlyStoredSuffix;
+  bool hasSpecialPointer;
+  int id;
+  unsigned long address;
+  union {
+    Node<DIM>* subnode;
+    const unsigned long* suffixStartBlock;
+    unsigned int suffixStartBlockIndex;
+    unsigned long suffix;
+    uintptr_t specialPointer;
+  };
 
-	const unsigned long* getSuffixStartBlock() const {
-		assert (exists && !hasSubnode);
-		return (directlyStoredSuffix)? &suffix : suffixStartBlock;
-	}
+  const unsigned long* getSuffixStartBlock() const {
+    assert (exists && !hasSubnode);
+    return (directlyStoredSuffix) ? &suffix : suffixStartBlock;
+  }
 };
-
 
 #endif /* NODEADDRESSCONTENT_H_ */
